Deduplicating test; moving json validation utils to utils.
